+++ wouldyourather/wouldyourather/SignInViewController.swift
@@ -1,0 +1,51 @@
+//
+//  SignInViewController.swift
+//  wouldyourather
+//
+//  Created by Matthew Nguyen on 11/11/22.
+//
+
+import UIKit
+import Parse
+
+class SignInViewController: UIViewController {
+
+    
+    
+    @IBOutlet weak var usernameField: UITextField!
+    @IBOutlet weak var passwordField: UITextField!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+    @IBAction func onSignIn(_ sender: Any) {
+        let username = usernameField.text!
+        let password = passwordField.text!
+        
+        PFUser.logInWithUsername(inBackground: username, password: password) { (user, error) in
+            if user != nil {
+                self.performSegue(withIdentifier: "loginSegue", sender: nil)
+            }
+            else {
+                print("Error: \(error?.localizedDescription ?? "sign in broke")")
+            }
+            
+        }
+    }
+    
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
